auth/djangoRest/helloDjangoRest/requirements.txt
Django==3.1.3
djangorestframework

